content/requirements.txt
pika==1.3.2
pyClamd==0.4.0
boto3==1.34.55
python-dotenv